+++ aim/web/ui/src/components/MediaPanel/MediaPanel.tsx
@@ -11,14 +11,16 @@ import { throttle } from 'components/Table/utils';
 
 import { ResizeModeEnum } from 'config/enums/tableEnums';
 import {
-  batchSendDelay,
-  imageFixedHeight,
-} from 'config/imagesConfigs/imagesConfig';
+  AUDIO_FIXED_HEIGHT,
+  BATCH_SEND_DELAY,
+  IMAGE_FIXED_HEIGHT,
+} from 'config/mediaConfigs/mediaConfigs';
 import { MediaItemAlignmentEnum } from 'config/enums/imageEnums';
 
 import blobsURIModel from 'services/models/media/blobsURIModel';
 
 import { IMediaPanelProps } from './MediaPanel.d';
+import { MediaTypeEnum } from './config';
 
 import './MediaPanel.scss';
 
@@ -75,8 +77,8 @@ function MediaPanel({
           blobUriArray.current = [];
         });
       }
-    }, batchSendDelay);
-  }, batchSendDelay);
+    }, BATCH_SEND_DELAY);
+  }, BATCH_SEND_DELAY);
 
   function onListScroll({ scrollOffset }: { scrollOffset: number }): void {
     if (Math.abs(scrollOffset - scrollTopOffset.current) > window.innerHeight) {
@@ -191,6 +193,25 @@ function MediaPanel({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
+  const mediaItemHeight = React.useMemo(() => {
+    if (additionalProperties?.alignmentType === MediaItemAlignmentEnum.Height) {
+      return (wrapperOffsetHeight * additionalProperties?.mediaItemSize) / 100;
+    } else if (
+      additionalProperties?.alignmentType === MediaItemAlignmentEnum.Width
+    ) {
+      return (wrapperOffsetWidth * additionalProperties?.mediaItemSize) / 100;
+    } else {
+      return mediaType === MediaTypeEnum.AUDIO
+        ? AUDIO_FIXED_HEIGHT
+        : IMAGE_FIXED_HEIGHT;
+    }
+  }, [
+    additionalProperties,
+    mediaType,
+    wrapperOffsetHeight,
+    wrapperOffsetWidth,
+  ]);
+
   return (
     <BusyLoaderWrapper
       isLoading={isLoading}
@@ -233,19 +254,7 @@ function MediaPanel({
                     sortFieldsDict={sortFieldsDict}
                     wrapperOffsetHeight={wrapperOffsetHeight - 48}
                     wrapperOffsetWidth={wrapperOffsetWidth}
-                    mediaItemHeight={
-                      additionalProperties?.alignmentType ===
-                      MediaItemAlignmentEnum.Height
-                        ? (wrapperOffsetHeight *
-                            additionalProperties?.mediaItemSize) /
-                          100
-                        : additionalProperties?.alignmentType ===
-                          MediaItemAlignmentEnum.Width
-                        ? (wrapperOffsetWidth *
-                            additionalProperties?.mediaItemSize) /
-                          100
-                        : imageFixedHeight
-                    }
+                    mediaItemHeight={mediaItemHeight}
                     focusedState={focusedState}
                     syncHoverState={syncHoverState}
                     orderedMap={orderedMap}