+++ aim/web/ui/src/components/kit/Icon/Icon.d.ts
@@ -101,4 +101,8 @@ export type IconName =
   | 'scatterplot'
   | 'y-axis'
   | 'image-explorer'
-  | 'image-properties';
+  | 'image-properties'
+  | 'play'
+  | 'pause'
+  | 'voice-on'
+  | 'voice-off';