+++ aim/web/ui/src/pages/RunDetail/TextsVisualizer/index.ts
@@ -1,0 +1,3 @@
+import TextsVisualizer from './TextsVisualizer';
+
+export default TextsVisualizer;