+++ aim/web/ui/src/pages/RunDetail/TraceVisualizationContainer/TraceVisualizationContainer.tsx
@@ -1,4 +1,4 @@
-import React, { memo, useEffect } from 'react';
+import React from 'react';
 
 import Menu from 'components/kit/Menu/Menu';
 import BusyLoaderWrapper from 'components/BusyLoaderWrapper/BusyLoaderWrapper';
@@ -6,23 +6,27 @@ import BusyLoaderWrapper from 'components/BusyLoaderWrapper/BusyLoaderWrapper';
 import useModel from 'hooks/model/useModel';
 
 import runTracesModel from 'services/models/runs/runTracesModel';
+import * as analytics from 'services/analytics';
+import { VisualizationMenuTitles } from 'services/models/runs/util';
 
 import DistributionsVisualizer from '../DistributionsVisualizer';
+import TextsVisualizer from '../TextsVisualizer';
 import ImagesVisualizer from '../ImagesVisualizer';
 import PlotlyVisualizer from '../PlotlyVisualizer';
 import { ITraceVisualizationContainerProps } from '../types';
+import AudiosVisualizer from '../AudiosVisualizer';
 
 import RangePanel from './RangePanel';
-import widthEmptyTraceCheck from './widthEmptyTraceCheck';
+import withEmptyTraceCheck from './withEmptyTraceCheck';
 
 import './TraceVisualizationContainer.scss';
 
 const traceTypeVisualization = {
   images: ImagesVisualizer,
   distributions: DistributionsVisualizer,
-  audios: () => null,
+  audios: AudiosVisualizer,
   videos: () => null,
-  texts: () => null,
+  texts: TextsVisualizer,
   figures: PlotlyVisualizer,
 };
 
@@ -33,7 +37,7 @@ function TraceVisualizationContainer({
   runParams,
 }: ITraceVisualizationContainerProps) {
   const runTracesModelData = useModel(runTracesModel);
-  useEffect(() => {
+  React.useEffect(() => {
     runTracesModel.initialize(
       runHash,
       traceType,
@@ -47,6 +51,10 @@ function TraceVisualizationContainer({
   }, [runHash, traceInfo, traceType]);
 
   const Visualizer = traceTypeVisualization[traceType];
+
+  React.useEffect(() => {
+    analytics.pageView(`[RunDetail] [${VisualizationMenuTitles[traceType]}]`);
+  }, [traceType]);
 
   return (
     <div className='TraceVisualizationWrapper'>
@@ -106,4 +114,4 @@ function TraceVisualizationContainer({
 
 TraceVisualizationContainer.displayName = 'TraceVisualizationContainer';
 
-export default memo(widthEmptyTraceCheck(TraceVisualizationContainer));
+export default React.memo(withEmptyTraceCheck(TraceVisualizationContainer));