+++ aim/web/ui/src/services/models/runs/settings.ts
@@ -2,8 +2,10 @@ import { QueryData } from './types';
 import {
   processDistributionsData,
   processImagesData,
+  processTextsData,
   processPlotlyData,
   reformatArrayQueries,
+  processAudiosData,
 } from './util';
 
 type InputItem = {
@@ -74,6 +76,64 @@ const settings: Record<string, SettingItem> = {
       },
     },
   },
+  audios: {
+    dataProcessor: processAudiosData,
+    sliders: {
+      record_range: {
+        defaultValue: [0, 50],
+        tooltip: 'Training step. Increments every time track() is called',
+        title: 'Steps',
+        sliderType: 'range',
+      },
+      index_range: {
+        defaultValue: [0, 50],
+        tooltip: 'Index in the list of audios passed to track() call',
+        title: 'Indices',
+        sliderType: 'range',
+      },
+    },
+    inputs: {
+      record_density: {
+        defaultValue: 50,
+        title: 'Steps count',
+        tooltip: 'Number of steps to display',
+      },
+      index_density: {
+        defaultValue: 5,
+        title: 'Indices count',
+        tooltip: 'Number of audios per step',
+      },
+    },
+  },
+  texts: {
+    dataProcessor: processTextsData,
+    sliders: {
+      record_range: {
+        defaultValue: [0, 50],
+        tooltip: 'Training step. Increments every time track() is called',
+        title: 'Steps',
+        sliderType: 'range',
+      },
+      index_range: {
+        defaultValue: [0, 50],
+        tooltip: 'Index in the list of texts passed to track() call',
+        title: 'Indices',
+        sliderType: 'range',
+      },
+    },
+    inputs: {
+      record_density: {
+        defaultValue: 50,
+        title: 'Steps count',
+        tooltip: 'Number of steps to display',
+      },
+      index_density: {
+        defaultValue: 5,
+        title: 'Indices count',
+        tooltip: 'Number of texts per step',
+      },
+    },
+  },
   figures: {
     dataProcessor: processPlotlyData,
     paramsToApi: (queryData?: QueryData) => {
@@ -97,7 +157,7 @@ const settings: Record<string, SettingItem> = {
     inputs: {
       record_range: {
         defaultValue: 1,
-        tooltip: 'Training step. To see Plotly tracked in the step.',
+        tooltip: 'Training step. To see figures tracked in the step.',
         title: 'Step',
       },
     },